move options to another file in config folder
