+++ lua/config/options.lua
